pyproject.toml
[tool.poetry]
name = "deep-research-agent-langgraph"
version = "0.1.0"
description = "LangGraph搭載 自律型リサーチエージェント"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = ">=0.1.0"
langgraph = ">=0.2.0"
langchain-openai = ">=0.1.0"
langchain-community = ">=0.0.20"
tavily-python = ">=0.3.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
requests = ">=2.31.0"
beautifulsoup4 = ">=4.12.0"
tenacity = ">=8.2.0"
python-dotenv = ">=1.0.0"
redis = {version = ">=5.0.0", optional = true}
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"
pytest-mock = ">=3.12.0"
pytest-asyncio = ">=0.21.0"
black = ">=23.11.0"
flake8 = ">=6.1.0"
mypy = ">=1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"









